[72048132] Define public APIs for adopters - combine IApplicationDelegate and ApplicationDelegate to AbstractApplicationDelegate
