+++ org.cloudfoundry.ide.eclipse.server.core/src/org/cloudfoundry/ide/eclipse/server/core/AbstractApplicationDelegate.java
@@ -16,15 +16,29 @@
  *  
  *  Contributors:
  *     Pivotal Software, Inc. - initial API and implementation
+ *     IBM Corporation - combine IApplicationDelegate and ApplicationDelegate
  ********************************************************************************/
 package org.cloudfoundry.ide.eclipse.server.core;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+
 import org.cloudfoundry.client.lib.archive.ApplicationArchive;
+import org.cloudfoundry.client.lib.domain.CloudApplication;
+import org.cloudfoundry.client.lib.domain.CloudService;
+import org.cloudfoundry.ide.eclipse.server.core.internal.CloudFoundryPlugin;
 import org.cloudfoundry.ide.eclipse.server.core.internal.CloudFoundryServer;
+import org.cloudfoundry.ide.eclipse.server.core.internal.CloudUtil;
+import org.cloudfoundry.ide.eclipse.server.core.internal.ValueValidationUtil;
+import org.cloudfoundry.ide.eclipse.server.core.internal.application.EnvironmentVariable;
 import org.cloudfoundry.ide.eclipse.server.core.internal.client.CloudFoundryApplicationModule;
+import org.cloudfoundry.ide.eclipse.server.core.internal.client.LocalCloudService;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.wst.server.core.IModule;
 import org.eclipse.wst.server.core.model.IModuleResource;
 
@@ -46,7 +60,7 @@ import org.eclipse.wst.server.core.model.IModuleResource;
  * </ul>
  * 
  */
-public interface IApplicationDelegate {
+public abstract class AbstractApplicationDelegate {
 
 	/**
 	 * 
@@ -55,7 +69,7 @@ public interface IApplicationDelegate {
 	 * @return true if the application requires that a URL be set when
 	 * publishing the application. False otherwise.
 	 */
-	public boolean requiresURL();
+	public abstract boolean requiresURL();
 
 	/**
 	 * A light-weight way of telling the framework whether this application
@@ -68,7 +82,7 @@ public interface IApplicationDelegate {
 	 * @return true if this delegate provides its own application serialisation
 	 * mechanism, or false otherwise
 	 */
-	public boolean providesApplicationArchive(IModule module);
+	public abstract boolean providesApplicationArchive(IModule module);
 
 	/**
 	 * An application archive generates input streams for an application's files
@@ -114,7 +128,7 @@ public interface IApplicationDelegate {
 	 * @throws CoreException if the application delegate provides an application
 	 * archive but it failed to create one.
 	 */
-	public ApplicationArchive getApplicationArchive(CloudFoundryApplicationModule module,
+	public abstract ApplicationArchive getApplicationArchive(CloudFoundryApplicationModule module,
 			CloudFoundryServer cloudServer, IModuleResource[] moduleResources, IProgressMonitor monitor) throws CoreException;
 
 	/**
@@ -123,7 +137,9 @@ public interface IApplicationDelegate {
 	 * @param deploymentInfo
 	 * @return non-null status.
 	 */
-	public IStatus validateDeploymentInfo(ApplicationDeploymentInfo deploymentInfo);
+	public IStatus validateDeploymentInfo(ApplicationDeploymentInfo deploymentInfo) {
+		return basicValidateDeploymentInfo(deploymentInfo);
+	}
 
 	/**
 	 * Resolve an application deployment for the given application. Return null
@@ -134,8 +150,10 @@ public interface IApplicationDelegate {
 	 * @return A new copy of the deployment information for an existing
 	 * application, or null if it cannot be resolved.
 	 */
-	public ApplicationDeploymentInfo resolveApplicationDeploymentInfo(CloudFoundryApplicationModule module,
-			CloudFoundryServer cloudServer);
+	public ApplicationDeploymentInfo resolveApplicationDeploymentInfo(CloudFoundryApplicationModule appModule,
+			CloudFoundryServer cloudServer) {
+		return parseApplicationDeploymentInfo(appModule.getApplication());
+	}
 
 	/**
 	 * Get a default application deployment information, regardless of whether
@@ -146,7 +164,99 @@ public interface IApplicationDelegate {
 	 * @param cloudServer
 	 * @return Non-null application deployment information with default values.
 	 */
+	/*
+	 * (non-Javadoc)
+	 * 
+	 * @see org.cloudfoundry.ide.eclipse.server.core.internal.application.
+	 * AbstractApplicationDelegate
+	 * #getDefaultApplicationDeploymentInfo(org.cloudfoundry.
+	 * ide.eclipse.internal.server.core.client.CloudFoundryApplicationModule,
+	 * org.cloudfoundry.ide.eclipse.server.core.internal.CloudFoundryServer,
+	 * org.eclipse.core.runtime.IProgressMonitor)
+	 */
 	public ApplicationDeploymentInfo getDefaultApplicationDeploymentInfo(CloudFoundryApplicationModule appModule,
-			CloudFoundryServer cloudServer, IProgressMonitor monitor) throws CoreException;
+			CloudFoundryServer cloudServer, IProgressMonitor monitor) throws CoreException {
+
+		// Set default values.
+		String appName = appModule.getDeployedApplicationName();
+		ApplicationDeploymentInfo deploymentInfo = new ApplicationDeploymentInfo(appName);
+		deploymentInfo.setMemory(CloudUtil.DEFAULT_MEMORY);
+
+		return deploymentInfo;
+	}
+
+	public static IStatus basicValidateDeploymentInfo(ApplicationDeploymentInfo deploymentInfo) {
+		IStatus status = Status.OK_STATUS;
+
+		String errorMessage = null;
+
+		if (deploymentInfo == null) {
+			errorMessage = "Missing application deployment information.";
+		}
+		else if (ValueValidationUtil.isEmpty(deploymentInfo.getDeploymentName())) {
+			errorMessage = "Missing application name in application deployment information.";
+		}
+		else if (deploymentInfo.getMemory() <= 0) {
+			errorMessage = "No memory set in application deployment information.";
+		}
+
+		if (errorMessage != null) {
+			status = CloudFoundryPlugin.getErrorStatus(errorMessage);
+		}
+
+		return status;
+	}
+
+	/**
+	 * Parses deployment information from a deployed Cloud Application. Returns
+	 * null if the cloud application is null.
+	 * @param cloudApplication deployed in a CF server
+	 * @return Parsed deployment information, or null if Cloud Application is
+	 * null.
+	 */
+	public static ApplicationDeploymentInfo parseApplicationDeploymentInfo(CloudApplication cloudApplication) {
+
+		if (cloudApplication != null) {
+
+			String deploymentName = cloudApplication.getName();
+			ApplicationDeploymentInfo deploymentInfo = new ApplicationDeploymentInfo(deploymentName);
+
+			deploymentInfo.setStaging(cloudApplication.getStaging());
+			deploymentInfo.setMemory(cloudApplication.getMemory());
+			List<String> boundServiceNames = cloudApplication.getServices();
+			if (boundServiceNames != null) {
+				List<CloudService> services = new ArrayList<CloudService>();
+				for (String name : boundServiceNames) {
+					if (name != null) {
+						services.add(new LocalCloudService(name));
+					}
+				}
+				deploymentInfo.setServices(services);
+			}
+
+			if (cloudApplication.getUris() != null) {
+				deploymentInfo.setUris(new ArrayList<String>(cloudApplication.getUris()));
+			}
+
+			Map<String, String> envMap = cloudApplication.getEnvAsMap();
+
+			if (envMap != null) {
+				List<EnvironmentVariable> variables = new ArrayList<EnvironmentVariable>();
+				for (Entry<String, String> entry : envMap.entrySet()) {
+					String varName = entry.getKey();
+					if (varName != null) {
+						EnvironmentVariable variable = new EnvironmentVariable();
+						variable.setVariable(varName);
+						variable.setValue(entry.getValue());
+						variables.add(variable);
+					}
+				}
+				deploymentInfo.setEnvVariables(variables);
+			}
+			return deploymentInfo;
+
+		}
+		return null;
+	}
 
 }
